+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/chassis/localizers/ThreeDeadWheelLocalizer.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.chassis;
+package org.firstinspires.ftc.teamcode.chassis.localizers;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.DualNum;
