+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/ActionCommand.java
@@ -1,0 +1,40 @@
+package org.firstinspires.ftc.teamcode.util;
+
+import com.acmerobotics.dashboard.FtcDashboard;
+import com.acmerobotics.dashboard.telemetry.TelemetryPacket;
+import com.acmerobotics.roadrunner.*;
+import com.arcrobotics.ftclib.command.Command;
+import com.arcrobotics.ftclib.command.Subsystem;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
+public class ActionCommand implements Command {
+    private final Action action;
+    private final Set<Subsystem> requirements = new HashSet<>();
+    private boolean finished = false;
+
+    public ActionCommand(Action action, Subsystem... requirements) {
+        this.action = action;
+        this.requirements.addAll(Arrays.asList(requirements));
+    }
+
+    @Override
+    public Set<Subsystem> getRequirements() {
+        return requirements;
+    }
+
+    @Override
+    public void execute() {
+        TelemetryPacket packet = new TelemetryPacket();
+        action.preview(packet.fieldOverlay());
+        finished = !action.run(packet);
+        FtcDashboard.getInstance().sendTelemetryPacket(packet);
+    }
+
+    @Override
+    public boolean isFinished() {
+        return finished;
+    }
+}
